protection for runtime updates

+++ git.sh
@@ -47,24 +47,28 @@ if [[ "$READBACK" == *"error"* ]]; then
 fi
 
 if [[ "$FIXIT" = "1" ]]; then
-  sudo su -c "sudo screen -X -S minew quit"
-  sudo su -c "sudo screen -X -S fakescreen quit"
-  sudo su -c "screen -ls minew | grep -E '\s+[0-9]+\.' | awk -F ' ' '{print $1}' | while read s; do screen -XS $s quit; done"
-  sudo su minerstat -c "screen -X -S fakescreen quit"
-  screen -ls minerstat-console | grep -E '\s+[0-9]+\.' | awk -F ' ' '{print $1}' | while read s; do screen -XS $s quit; done
-  sudo su minerstat -c "screen -ls minerstat-console | grep -E '\s+[0-9]+\.' | awk -F ' ' '{print $1}' | while read s; do screen -XS $s quit; done"
-  sudo killall node
-  sleep 0.25
-  cd /home/minerstat
-  sudo rm /home/minerstat/recovery.sh
-  wget -o /dev/null https://labs.minerstat.farm/repo/minerstat-os/-/raw/master/core/recovery.sh
-  sudo chmod 777 /home/minerstat/recovery.sh
-  sudo bash /home/minerstat/recovery.sh
-  sudo bash /home/minerstat/minerstat-os/bin/overclock.sh &
-  sleep 15
-  sudo su minerstat -c "screen -A -m -d -S fakescreen sh /home/minerstat/minerstat-os/bin/fakescreen.sh"
-  sleep 2
-  sudo su minerstat -c "screen -A -m -d -S minerstat-console sudo /home/minerstat/minerstat-os/launcher.sh"
+  if [[ -z $1 ]]; then
+    sudo su -c "sudo screen -X -S minew quit"
+    sudo su -c "sudo screen -X -S fakescreen quit"
+    sudo su -c "screen -ls minew | grep -E '\s+[0-9]+\.' | awk -F ' ' '{print $1}' | while read s; do screen -XS $s quit; done"
+    sudo su minerstat -c "screen -X -S fakescreen quit"
+    screen -ls minerstat-console | grep -E '\s+[0-9]+\.' | awk -F ' ' '{print $1}' | while read s; do screen -XS $s quit; done
+    sudo su minerstat -c "screen -ls minerstat-console | grep -E '\s+[0-9]+\.' | awk -F ' ' '{print $1}' | while read s; do screen -XS $s quit; done"
+    sudo killall node
+    sleep 0.25
+    cd /home/minerstat
+    sudo rm /home/minerstat/recovery.sh
+    wget -o /dev/null https://labs.minerstat.farm/repo/minerstat-os/-/raw/master/core/recovery.sh
+    sudo chmod 777 /home/minerstat/recovery.sh
+    sudo bash /home/minerstat/recovery.sh
+    sudo bash /home/minerstat/minerstat-os/bin/overclock.sh &
+    sleep 15
+    sudo su minerstat -c "screen -A -m -d -S fakescreen sh /home/minerstat/minerstat-os/bin/fakescreen.sh"
+    sleep 2
+    sudo su minerstat -c "screen -A -m -d -S minerstat-console sudo /home/minerstat/minerstat-os/launcher.sh"
+  else
+    echo "File system corrupted but not attempting to auto-fix because not manually requested or the machine not just booting"
+  fi
 fi
 
 #sudo rm /home/minerstat/debug.txt
